pyproject.toml
[project]
name = "confidoc-backend"
version = "0.1.0"
description = "Backend SaaS de confidentialité documentaire pour professions réglementées"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "ConfiDoc", email = "dev@confidoc.fr" }]

dependencies = [
    # Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",

    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.1.0",
    "cryptography>=42.0.0",
    "python-multipart>=0.0.7",

    # Async tasks
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",

    # Storage
    "minio>=7.2.0",

    # Logging & Monitoring
    "structlog>=24.1.0",

    # Rate limiting
    "slowapi>=0.1.9",

    # Utils
    "httpx>=0.27.0",
    "python-magic>=0.4.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

processing = [
    # OCR
    "pytesseract>=0.3.10",
    "Pillow>=10.2.0",
    "pdf2image>=1.17.0",
    # PDF Parsing
    "PyMuPDF>=1.23.0",
    # NER
    "spacy>=3.7.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
